# personal-website-admin/src/main/java/com/jinlong/personalwebsitesys/admin/domain/dto/PasswordUpdateRequest.java
package com.jinlong.personalwebsitesys.admin.domain.dto;

/** 密码请求不自动生成 toString，避免密码进入日志。 */
public class PasswordUpdateRequest {
    private String currentPassword;
    private String newPassword;
    private String confirmPassword;

    public String getCurrentPassword() { return currentPassword; }
    public void setCurrentPassword(String value) { currentPassword = value; }
    public String getNewPassword() { return newPassword; }
    public void setNewPassword(String value) { newPassword = value; }
    public String getConfirmPassword() { return confirmPassword; }
    public void setConfirmPassword(String value) { confirmPassword = value; }
}